# openapi/notebook-api.yaml
openapi: 3.1.0
info:
  title: Notebook API
  description: |
    REST API for managing notebooks.
    This OpenAPI document is the source of truth for the Notebook REST API.
    Java API interfaces and DTOs are generated from this contract.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
tags:
  - name: Notebooks
    description: Notebook management operations
  - name: NotebookIntelligence
    description: Search and retrieve MongoDB-enriched notebook information
paths:
  /api/v1/notebooks:
    get:
      tags:
        - Notebooks
      summary: Get all notebooks
      description: Returns all notebooks.
      operationId: getAllNotebooks
      responses:
        '200':
          description: Successfully retrieved notebooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notebook'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Notebooks
      summary: Create a notebook
      description: Creates a new notebook.
      operationId: createNotebook
      requestBody:
        required: true
        description: Notebook to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotebookRequest'
      responses:
        '201':
          description: Notebook successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notebook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/notebooks/{id}:
    get:
      tags:
        - Notebooks
      summary: Get notebook by ID
      operationId: getNotebookById
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      responses:
        '200':
          description: Notebook found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notebook'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Notebooks
      summary: Update notebook
      operationId: updateNotebook
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      requestBody:
        required: true
        description: Updated notebook
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotebookRequest'
      responses:
        '200':
          description: Notebook successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notebook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Notebooks
      summary: Delete notebook
      operationId: deleteNotebook
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      responses:
        '204':
          description: Notebook successfully deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/intelligence/notebooks:
    servers:
      - url: http://localhost:8082
        description: Local Notebook Intelligence Service

    get:
      tags:
        - NotebookIntelligence
      summary: Search enriched notebooks
      description: >
        Search notebook intelligence documents by title.
        Omit q to return all enriched notebooks.
      operationId: searchNotebookIntelligence
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive title search
          schema:
            type: string
      responses:
        '200':
          description: Matching enriched notebooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotebookIntelligence'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/intelligence/notebooks/{id}:
    servers:
      - url: http://localhost:8082
        description: Local Notebook Intelligence Service

    get:
      tags:
        - NotebookIntelligence
      summary: Get enriched notebook by ID
      operationId: getNotebookIntelligenceById
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      responses:
        '200':
          description: Enriched notebook found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotebookIntelligence'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/intelligence/notebooks/topic/{topic}:
    servers:
      - url: http://localhost:8082
        description: Local Notebook Intelligence Service

    get:
      tags:
        - NotebookIntelligence
      summary: Find enriched notebooks by topic
      operationId: getNotebooksByTopic
      parameters:
        - name: topic
          in: path
          required: true
          description: Topic to match, for example CLOUD NATIVE
          schema:
            type: string
      responses:
        '200':
          description: Notebooks matching the topic
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotebookIntelligence'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    NotebookId:
      name: id
      in: path
      required: true
      description: Unique identifier of the notebook
      schema:
        type: string
        format: uuid
  schemas:
    Notebook:
      type: object
      required:
        - id
        - title
        - description
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: Unique notebook identifier
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Notebook title
        description:
          type:
            - string
          maxLength: 5000
          description: Optional notebook description
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: Notebook creation timestamp
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Last notebook modification timestamp
    CreateNotebookRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Notebook title
        description:
          type:
            - string
            - 'null'
          maxLength: 5000
          description: Optional notebook description
    UpdateNotebookRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Notebook title
        description:
          type:
            - string
            - 'null'
          maxLength: 5000
          description: Optional notebook description
    ProblemDetail:
      type: object
      required:
        - status
        - title
        - detail
        - timestamp
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code
        title:
          type: string
          description: Short description of the error
        detail:
          type: string
          description: Detailed error description
        instance:
          type: string
          description: Request URI
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          description: Distributed tracing identifier

    NotebookIntelligence:
      type: object
      required:
        - notebookId
        - title
        - status
        - topics
        - keywords
        - entities
        - createdAt
        - updatedAt
      properties:
        notebookId:
          type: string
          format: uuid
          description: ID of the original PostgreSQL notebook
        title:
          type: string
          maxLength: 255
        description:
          type: [string, 'null']
          maxLength: 5000
        status:
          type: string
          description: Enrichment processing status
          enum:
            - PENDING
            - PROCESSING
            - COMPLETED
            - FAILED
        topics:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        entities:
          type: array
          items:
            $ref: '#/components/schemas/ExtractedEntity'
        summary:
          type: [string, 'null']
        source:
          $ref: '#/components/schemas/IntelligenceSourceEvent'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        enrichedAt:
          type: [string, 'null']
          format: date-time

    ExtractedEntity:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string

    IntelligenceSourceEvent:
      type: object
      properties:
        eventId:
          type: [string, 'null']
        eventType:
          type: string
        topic:
          type: string
        partition:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        schemaVersion:
          type: integer
          format: int32
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Notebook not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'